club entity 생성, student에 매핑
