+++ src/main/java/live/goapi/domain/club/entity/Club.java
@@ -1,0 +1,28 @@
+package live.goapi.domain.club.entity;
+
+import live.goapi.domain.student.entity.Student;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity @Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Club {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "club_id")
+    private Long clubId;
+
+    @Column(name = "club_name")
+    private String clubName;
+
+    @OneToMany
+    private List<Student> students;
+}
