--- src/main/java/live/goapi/domain/gsmapi/student/exception/NotFoundStudentException.java
@@ -1,16 +1,0 @@
-package live.goapi.domain.gsmapi.student.exception;
-
-import live.goapi.global.exception.ErrorCode;
-import live.goapi.global.exception.collection.BasicException;
-import lombok.Getter;
-
-@Getter
-public class NotFoundStudentException extends BasicException {
-
-    private final ErrorCode errorCode;
-
-    public NotFoundStudentException(String message) {
-        super(message);
-        errorCode = ErrorCode.TEACHER_NOT_FOUND;
-    }
-}
