--- src/main/java/live/goapi/domain/gsmapi/student/presentation/dto/response/ResponseStudent.java
@@ -1,16 +1,0 @@
-package live.goapi.domain.gsmapi.student.presentation.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ResponseStudent {
-    private String studentName;
-    private String studentNumber;
-    private String studentMajor;
-}
