--- src/main/java/live/goapi/domain/gsmapi/teacher/exception/NotFoundTeacherException.java
@@ -1,15 +1,0 @@
-package live.goapi.domain.gsmapi.teacher.exception;
-
-import live.goapi.global.exception.ErrorCode;
-import live.goapi.global.exception.collection.BasicException;
-import lombok.Getter;
-
-@Getter
-public class NotFoundTeacherException extends BasicException {
-    private final ErrorCode errorCode;
-
-    public NotFoundTeacherException(String message) {
-        super(message);
-        errorCode = ErrorCode.TEACHER_NOT_FOUND;
-    }
-}
