--- src/main/java/live/goapi/domain/gsmapi/teacher/presentation/dto/response/ResponseTeacher.java
@@ -1,11 +1,0 @@
-package live.goapi.domain.gsmapi.teacher.presentation.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ResponseTeacher {
-    private String teacherName;
-    private String subject;
-}
