+++ src/main/java/live/goapi/domain/teacher/presentation/dto/response/ResponseTeacher.java
@@ -1,0 +1,11 @@
+package live.goapi.domain.teacher.presentation.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ResponseTeacher {
+    private String teacherName;
+    private String subject;
+}
