+++ src/main/java/live/goapi/global/exception/handler/GlobalExceptionHandler.java
@@ -1,10 +1,10 @@
 package live.goapi.global.exception.handler;
 
-import live.goapi.domain.gsmapi.student.exception.NotFoundStudentException;
+import live.goapi.domain.student.exception.NotFoundStudentException;
 import live.goapi.domain.member.exception.MemberNotFoundException;
 import live.goapi.domain.auth.exception.PasswordMismatchException;
 import live.goapi.global.exception.ErrorResponse;
-import live.goapi.domain.gsmapi.teacher.exception.NotFoundTeacherException;
+import live.goapi.domain.teacher.exception.NotFoundTeacherException;
 import live.goapi.global.security.exception.TokenExpirationException;
 import live.goapi.global.security.exception.TokenNotValidException;
 import lombok.extern.slf4j.Slf4j;
